Contains a trial website for End project
This is the animated header that I had created for the use of my final project however during the creation of the website it did not follow my logo and colour theme. This caused a problem and a clash of concepts, as my end project was quite cheerful and cartoon(y) which did not follow the realistic, professional banner I had created.